Extract patterns and corresponding methods from MXSession to a dedicated MXPatterns class It fixes issue on unitary test, because MXSession cannot be loaded in UT due to Olm lib
